[effects] Rewrite the Fade Desktop effect
The main motive for rewriting this effect back in C++ was to fix bug
321201. C++ had been picked because scripted effects can't do full
screen effects.

The only difference between the old effect and the rewritten effect is
that the latter doesn't fade moving windows. Other than that, they are
pretty similar.

=== Demos

Switching between virtual desktops:
{F6128539}

Moving a window to another virtual desktop:
{F6128541}

Switching to another virtual desktop while previous transition(s) is(are)
still in progress:
{F6128547}

Before: the Fade desktop effect is false-triggered when switching virtual
desktops with the Desktop Grid effect:
{F6128552}

NOTE: the Fade effect is disabled.

After: the Fade desktop effect is not triggered when switching virtual
desktops with the Desktop Grid effect (expected behavior):
{F6128559}

Before: the Fade desktop effect is false-triggered when switching
virtual desktops with the Desktop Cube effect:
{F6128569}

NOTE: the Fade effect is disabled.

After: the Fade desktop effect is not triggered when switching virtual
desktops with the Desktop Cube effect (expected behavior):
{F6128582}

BUG: 321201
FIXED-IN: 5.XX.0