[effects] Rewrite the Magic Lamp effect
Summary:
The current Magic lamp effect doesn't work correctly if a window is
slightly below its icon in the task manager or when the panel is "between"
screens(see bug 361121).

This effect is no longer one single animation. Now, it's composed of
multiple smaller animations, so called "stages". Each stage animates one
particular stage of the whole "magic lamp" animation:

* "Bump": this stage animates the raising of the window if it's below
  its icon;
* "Stretch1": this stage animates the stretching of the window;
* "Stretch2": this stage animates the stretching and the lowering of the
  window back to its original position. This animation will be activated
  when the window is unminimized and it's slightly below its icon in the
  task manager(if the panel is on the bottom screen edge);
* "Squash": this stage animates squashing of the window to the task
  manager.

If a window is minimized, then the following stages will animate it:

    Stretch1 -> Squash

or if it's slightly below its icon, then the following stages will
animate it:

    Bump -> Stretch1 -> Squash

If a window is unminimized, the the following stages will animate it:

    Squash -> Stretch1

or if it's slightly below its icon, then the following stages will
animate it:

    Squash -> Stretch2

Even though the "Bump" animation won't be activated when the window is
unminimized, one still can toggle direction of that animation by
unminimizing the window. One can do that because this effect allows
implicit transitions between state machines for the minimize and the
unminimize animation. That implicit transition is achieved by toggling
direction of the currently running animation(in slotWindowMinimized and
slotWindowUnminimized).

For shaping windows, we use QEasingCurve. This allows us to use
Bezier curves or easing curves(InOutQuad, InOutSine, InOutElastic, etc) to
control how windows are distorted. InOutSine is pretty good curve, so I
picked it.

Even though that's still WIP, the effect is pretty much usable right now.

TODO:
* pick good defaults;
* revisit easing curves;
* add a comment about how this effect works and what it's doing.

That's not 5.14 stuff.

Test Plan:
Test plan 1:
* Minimized/Unminimized System Settings windows;

Test plan 2:
* Moved System Settings window slightly below the task manager;
* Minimized/Unminimized the window.
* (during the minimize animation, the window had been bumped/raised);
* (during the unminimize animation, the window was expanding and lowering at the same time).

Test plan 3:
* (attached the second screen, put panel on the right edge of the left screen);
* Moved System Settings window to the left screen, minimized/uniminimized it;
* (during the magic lamp animation, didn't notice any artefacts);
* Moved System Settings window to the right screen, minimized/unminimized it;
* (during the magic lamp animation, didn't notice any artefacts).

Reviewers: #kwin, #plasma, #vdg

Subscribers: abetts, kwin

Tags: #kwin

Differential Revision: https://phabricator.kde.org/D14749