[wayland] Apply window rules only to xdg-shell clients
There are rules that have to be applied only once, e.g. every Remember
and Apply Initially rule, as well any rule that involves configuring
the client. In the best scenario the compositor would evaluate such rules
when the client is about to be mapped.

This change limits window rules only to xdg-shell clients because right
now only this protocol lets compositors intervene in the client
initialization process. Also, it makes things a bit easier for us on the
compositor side.

Though not all issues are addressed with xdg-shell protocol. There is no
way for a client to express its intent to be mapped. Thus if a client is
mapped more than once, then the compositor won't be able to re-evaluate
remember rules for example. In order to make this work we need changes
in the protocol.