[nightcolor] Use a dedicated class for detection of system clock changes
The ClockSkewNotifier provides a convenient way for monitoring system
clock changes. One of the key ideas was to hide platform details from
users of the class. This allows us to add a QTimer fallback path for
operating systems that don't have native support for detecting system
clock changes.

In long term, I would like to move the new class to Frameworks.